+++ Order/models.py
@@ -40,19 +40,19 @@ class Checkout(models.Model):
     client_type = models.CharField(choices = CLIENT_TYPE, max_length=50 , default = '' )
     payment_type = models.CharField(choices = PAYMENT_TYPE, max_length=50 , default = '' )
     
-    tip = models.PositiveBigIntegerField(default = 0)
+    tip = models.FloatField(default = 0)
 
     tax_applied = models.FloatField(default=0, verbose_name='Tax Applied in Percentage')
     tax_amount = models.FloatField(default=0, verbose_name='Tax total amount')
     
-    total_service_price = models.PositiveBigIntegerField(default = 0 , null=True, blank=True)
-    total_product_price = models.PositiveBigIntegerField(default = 0 , null=True, blank=True)
-    total_voucher_price = models.PositiveBigIntegerField(default = 0 , null=True, blank=True)
-    total_membership_price = models.PositiveBigIntegerField(default = 0 , null=True, blank=True)
+    total_service_price = models.FloatField(default = 0 , null=True, blank=True)
+    total_product_price = models.FloatField(default = 0 , null=True, blank=True)
+    total_voucher_price = models.FloatField(default = 0 , null=True, blank=True)
+    total_membership_price = models.FloatField(default = 0 , null=True, blank=True)
     
-    service_commission = models.PositiveBigIntegerField(default = 0 , null=True, blank=True)
-    product_commission = models.PositiveBigIntegerField(default = 0 , null=True, blank=True)
-    voucher_commission = models.PositiveBigIntegerField(default = 0 , null=True, blank=True)
+    service_commission = models.FloatField(default = 0 , null=True, blank=True)
+    product_commission = models.FloatField(default = 0 , null=True, blank=True)
+    voucher_commission = models.FloatField(default = 0 , null=True, blank=True)
     
     service_commission_type = models.CharField( max_length=50 , default = '')
     product_commission_type = models.CharField( max_length=50 , default = '')
@@ -143,15 +143,15 @@ class Order(models.Model):
     
     quantity = models.PositiveBigIntegerField(default= 0)
     
-    current_price =models.PositiveBigIntegerField(default = 0)
-    tip =models.PositiveBigIntegerField(default = 0)
-    gst = models.PositiveBigIntegerField(default = 0)
+    current_price =models.FloatField(default = 0)
+    tip =models.FloatField(default = 0)
+    gst = models.FloatField(default = 0)
     total_price = models.DecimalField(default = 0 , max_digits=10, decimal_places=5)
     sold_quantity = models.PositiveBigIntegerField(default = 0)
     
     discount_percentage = models.FloatField(default= 0)
     discount_price = models.FloatField(default= 0)
-    price = models.PositiveBigIntegerField(default= 0)
+    price = models.FloatField(default= 0)
 
     is_redeemed = models.BooleanField(default=False)
     redeemed_type = models.CharField(default='', max_length=300)
